[simplify] refactor: move reviewer identity formatting to the output layer
The config layer was building a display string that both output modules
imported back, inverting the usual direction. Move it next to statusLineText,
which already owns the sibling line, and have the console reporter import it
from there.

Name the active branch of the parse result instead of repeating the Extract
expression at three signatures, derive the overlay thinking budget from the
adapter defaults rather than restating the union, and import the shared
reviewer environment variable names in the end-to-end test instead of
re-typing them.
